+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - Renamed `AnyFromMap` field handler to `MapToAny`.
 - Added map conversions as part of converter handler chain so that map conversions can be done by default.
-- Added `MapConversions.toPropertiesMap(Object)` method that deep converts an object to a map of field names and values.
+- Added `MapConversions.toPropertiesMap(Object)` method that deep converts an object to a map of field names and values with support for cyclic references.
 - Renamed `Configuration.defaultConfiguration` to `Configuration.defaults`.
+- Added `ConversionContext` to track cyclic references during conversions.
 
 ---